attempt to generate dependencies
