+++ packages/riverpod_analysis/.gitignore
@@ -1,0 +1,6 @@
+# Files and directories created by pub.
+.dart_tool/
+.packages
+
+# Conventional directory for build output.
+build/
