+++ packages/riverpod_analysis/README.md
@@ -1,0 +1,2 @@
+A sample command-line application with an entrypoint in `bin/`, library code
+in `lib/`, and example unit test in `test/`.
