+++ packages/riverpod_analysis/analysis_options.yaml
@@ -1,0 +1,6 @@
+include: ../../analysis_options.yaml
+linter:
+  rules:
+    # Riverpod_lint has no public API
+    public_member_api_docs: false
+
