+++ packages/riverpod_analysis/lib/src/analyzer_utils.dart
