+++ packages/riverpod_analysis/lib/src/utils.dart
