+++ packages/riverpod_analysis/pubspec.yaml
@@ -1,0 +1,17 @@
+name: riverpod_analysis
+description: A sample command-line application.
+version: 1.0.0
+# repository: https://github.com/my_org/my_repo
+
+environment:
+  sdk: '>=2.18.0 <3.0.0'
+
+dependencies:
+  path: ^1.8.0
+  analyzer: ^4.0.0
+  collection: ^1.16.0
+  meta: ^1.7.0
+
+dev_dependencies:
+  lints: ^2.0.0
+  test: ^1.16.0
