+++ packages/riverpod_generator/pubspec.yaml
@@ -14,8 +14,10 @@ dependencies:
   meta: ^1.7.0
   riverpod_annotation: ^1.0.0
   source_gen: ^1.2.0
+  riverpod_analysis: any
 
 dev_dependencies:
   build_runner: ^2.1.7
   riverpod: ^2.0.0
   test: ^1.21.0
+  build_test: any
