+++ packages/riverpod_generator/pubspec_overrides.yaml
@@ -4,3 +4,5 @@ dependency_overrides:
     path: ../riverpod
   riverpod_annotation:
     path: ../riverpod_annotation
+  riverpod_analysis:
+    path: ../riverpod_analysis
